factor out the pool asset item to be reused
