+++ ui/app/src/components/poolList/PoolListItem.vue
@@ -4,6 +4,7 @@ import { defineComponent, PropType } from "vue";
 import { getAssetLabel, useAssetItem } from "@/components/shared/utils";
 import { LiquidityProvider, Pool } from "ui-core";
 import { useRouter } from 'vue-router';
+import AssetItemPool from "@/components/shared/AssetItemPool.vue";
 
 export default defineComponent({
   props: {
@@ -12,35 +13,18 @@ export default defineComponent({
     },
   },
 
+  components: {
+    AssetItemPool,
+  },
+
   setup(props) {
     const router = useRouter();
+    
     const fromSymbol = computed(() =>
       props.accountPool?.pool.amounts[1].asset
         ? getAssetLabel(props.accountPool?.pool.amounts[1].asset)
         : ""
     );
-    const fromAsset = useAssetItem(fromSymbol);
-    const fromToken = fromAsset.token;
-    const fromBackgroundStyle = fromAsset.background;
-    const fromTokenImage = computed(() => {
-      if (!fromToken.value) return "";
-      const t = fromToken.value;
-      return t.imageUrl;
-    });
-
-    const toSymbol = computed(() =>
-      props.accountPool?.pool.amounts[0].asset
-        ? getAssetLabel(props.accountPool?.pool.amounts[0].asset)
-        : ""
-    );
-    const toAsset = useAssetItem(toSymbol);
-    const toToken = toAsset.token;
-    const toBackgroundStyle = toAsset.background;
-    const toTokenImage = computed(() => {
-      if (!toToken.value) return "";
-      const t = toToken.value;
-      return t.imageUrl;
-    });
 
     const handleClick = () => {
       router.push(`/pool/${fromSymbol.value.toLowerCase()}`)
@@ -48,12 +32,6 @@ export default defineComponent({
 
     return {
       handleClick,
-      fromSymbol,
-      fromBackgroundStyle,
-      fromTokenImage,
-      toSymbol,
-      toBackgroundStyle,
-      toTokenImage,
     };
   },
 });
@@ -61,29 +39,9 @@ export default defineComponent({
 
 <template>
   <div class="pool-list-item" @click="handleClick">
-    <div class="image">
-      <img
-        v-if="fromTokenImage"
-        width="22"
-        height="22"
-        :src="fromTokenImage"
-        class="info-img"
-      />
-      <div class="placeholder" :style="fromBackgroundStyle" v-else></div>
-      <img
-        v-if="toTokenImage"
-        width="22"
-        height="22"
-        :src="toTokenImage"
-        class="info-img"
-      />
-      <div class="placeholder" :style="toBackgroundStyle" v-else></div>
-    </div>
-    <div class="symbol">
-      <span>{{ fromSymbol }}</span>
-      /
-      <span>{{ toSymbol }}</span>
-    </div>
+    <AssetItemPool 
+      :pool="accountPool"
+    />
     <div class="button">Manage</div>
   </div>
 </template>
@@ -100,34 +58,6 @@ export default defineComponent({
   &:hover {
     cursor: pointer;
     background: $c_gray_50;
-  }
-
-  .image {
-    height: 22px;
-
-    & > * {
-      border-radius: 16px;
-
-      &:nth-child(2) {
-        position: relative;
-        left: -6px;
-      }
-    }
-  }
-
-  .placeholder {
-    display: inline-block;
-    background: #aaa;
-    box-sizing: border-box;
-    border-radius: 16px;
-    height: 22px;
-    width: 22px;
-    text-align: center;
-  }
-
-  .symbol {
-    font-size: $fs_md;
-    color: $c_text;
   }
 
   .button {
